+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Added
 - Model Fields - Support "serialize" option for String type fields.
+- Pie Charts - Support Pie charts with a group by on an association field.
 
 ### Changed
 - Authentication - Add an option to set the model name for internal forest user authentication option.